+++ hydra-node/src/Hydra/Network/Heartbeat.hs
@@ -78,8 +78,6 @@ heartbeatDelay = 0.5
 livenessDelay :: DiffTime
 livenessDelay = 3
 
-type ConnectionMessages m = Connectivity -> m ()
-
 -- | Wrap a lower-level `NetworkComponent` and handle sending/receiving of heartbeats.
 --
 -- Note that the type of consumed and sent messages can be different.
@@ -89,35 +87,34 @@ withHeartbeat ::
   ) =>
   -- | This node's id, used to identify `Heartbeat` messages broadcast to peers.
   NodeId ->
-  -- | Callback listening to peers' status change as computed by the `withIncomingHeartbeat` layer.
-  ConnectionMessages m ->
   -- | Underlying `NetworkComponent` for sending and consuming `Heartbeat` messages.
   NetworkComponent m (Heartbeat inbound) (Heartbeat outbound) a ->
   -- | Returns a network component that can be used to send and consume arbitrary messages.
   -- This layer will take care of peeling out/wrapping messages into `Heartbeat`s.
-  NetworkComponent m inbound outbound a
-withHeartbeat nodeId connectionMessages withNetwork callback action = do
+  NetworkComponent m (Either Connectivity inbound) outbound a
+withHeartbeat nodeId withNetwork callback action = do
   heartbeat <- newTVarIO initialHeartbeatState
   lastSent <- newTVarIO Nothing
-  withNetwork (updateStateFromIncomingMessages heartbeat connectionMessages callback) $ \network ->
-    withAsync (checkRemoteParties heartbeat connectionMessages) $ \_ ->
+  withNetwork (updateStateFromIncomingMessages heartbeat callback) $ \network ->
+    withAsync (checkRemoteParties heartbeat onConnectivityChanged) $ \_ ->
       withAsync (checkHeartbeatState nodeId lastSent network) $ \_ ->
         action (updateStateFromOutgoingMessages nodeId lastSent network)
+ where
+  onConnectivityChanged = callback . Left
 
 updateStateFromIncomingMessages ::
   (MonadSTM m, MonadMonotonicTime m) =>
   TVar m HeartbeatState ->
-  ConnectionMessages m ->
-  NetworkCallback inbound m ->
+  NetworkCallback (Either Connectivity inbound) m ->
   NetworkCallback (Heartbeat inbound) m
-updateStateFromIncomingMessages heartbeatState connectionMessages callback = \case
-  Data nodeId msg -> notifyAlive nodeId >> callback msg
+updateStateFromIncomingMessages heartbeatState callback = \case
+  Data nodeId msg -> notifyAlive nodeId >> callback (Right msg)
   Ping nodeId -> notifyAlive nodeId
  where
   notifyAlive peer = do
     now <- getMonotonicTime
     aliveSet <- alive <$> readTVarIO heartbeatState
-    unless (peer `Map.member` aliveSet) $ connectionMessages (Connected peer)
+    unless (peer `Map.member` aliveSet) $ callback (Left $ Connected peer)
     atomically $
       modifyTVar' heartbeatState $ \s ->
         s
@@ -165,14 +162,14 @@ checkRemoteParties ::
   , MonadSTM m
   ) =>
   TVar m HeartbeatState ->
-  ConnectionMessages m ->
+  (Connectivity -> m ()) ->
   m ()
-checkRemoteParties heartbeatState connectionMessages =
+checkRemoteParties heartbeatState onConnectivity =
   forever $ do
     threadDelay (heartbeatDelay * 2)
     now <- getMonotonicTime
     updateSuspected heartbeatState now
-      >>= mapM_ (connectionMessages . Disconnected)
+      >>= mapM_ (onConnectivity . Disconnected)
 
 updateSuspected :: MonadSTM m => TVar m HeartbeatState -> Time -> m (Set NodeId)
 updateSuspected heartbeatState now =